--- Datas/MainFile.json
@@ -1,16 +1,0 @@
-{
-    "broken" : "破碎的",
-    "hello" : "你好",
-    "Unit 1" : {
-        "jungle" : "丛林",
-        "edge" : "边缘",
-        "pig" : "猪",
-        "Unit2": {
-            "apple" : "苹果",
-            "banana" : "香蕉",
-            "pig" : "猪",
-            "jungle" : "丛林",
-            "broken" : "破碎的"
-        }
-    }
-}
