+++ Readme.md
@@ -4,7 +4,7 @@
 
 The tester is a simple and weak word tester. But it will get stronger as I grow up!
 
-The program was made in Python 3.10.4 (Microsoft Store Edition).
+The program was made in Python 3.10.5 (Microsoft Store Edition).
 
 ## How to use
 
@@ -12,17 +12,17 @@ The program was made in Python 3.10.4 (Microsoft Store Edition).
 
 To run the program, input this piece of code into shells like, for example, `cmd` and `powershell`.
   ```powershell
-  python mainp.py functions.py
+  python main.py
   ```
 
-**<u>Warning: You must make sure the `mainp.py`, `functions.py`, and `Datas` folder are in the same path. Instead, the program may have unexcepted bugs!</u>**
+**<u>Warning: You must make sure these Python files and two Datas folder are in the same path. Instead, the program may have unexcepted bugs!</u>**
 
 - About the MainFile :
-  - The MainFile files (including `MainFile.txt` and `MainFile.json`) are used to store questions that you have been asked.And they must be in the "Datas" folder. Without these files, or if the file path has problems, the program may report an error (as the time used for development is so short, we can't make sure it will report an error 100%).
-  - The program has the JSON mode and the TXT mode. The txt mode uses the "MainFile.txt" file and the json mode uses the `MainFile.json` file. Editing the txt file is easy, but keeping the format right and streamlining is a bit difficult. Maybe the little format mistake in a superbly large txt file could make you superbly sad, and it is difficult to group words. As a result, I recommend that you use the json mode. Although editing a JSON file is difficult and troublesome, the JSON mode supports nested dictionaries, which the TXT mode cannot do.
+  - The Datas folders (including `TxtDatas` and `JsonDatas`) are used to store questions that you have been asked.And they must be in the "Datas" folder. Without these files, or if the file path has problems, the program may report an error (as the time used for development is so short, we can't make sure it will report an error 100%).
+  - The program has the JSON mode and the TXT mode. The txt mode uses the `TxtDatas/MainFile.txt` file and the json mode uses the `JsonDatas/MainFile.json` file. Editing the txt file is easy, but keeping the format right and streamlining is a bit difficult. Maybe the little format mistake in a superbly large txt file could make you superbly sad, and it is difficult to group words. As a result, I recommend that you use the json mode. Although editing a JSON file is difficult and troublesome, the JSON mode supports nested dictionaries, which the TXT mode cannot do.
   - <u>**Json mode warning: If a dictionary contains both words and other types of data, the program will automatically skip the other types of data and keep the dictionary when reading it.**</u>
 
-## The planning in the future
+## Plans in the future
 
 The project's development has been stopped as I have to study in school. But I still collect the advice and bugs of the program. If you discover any bugs or have any suggestions, I would be most welcome to open an issue!
