--- docs/Blogs.html
@@ -1,15 +1,0 @@
-<html>
-<head><title>Captain Wordtester Blog</title></head>
-<body>
-
-<h2># There are blogs about the Wordtester<h2>
-<br>
-<h3># 2022.6.6 17:14 </h3>
-Something I want to add in the Wordtester in futuer: <br>
-    1. We will add a "setting.json" in the ".\Datas",which can remember your personal settings of the programe.Yeah,I will add a new setting! <br>
-    2.A new setting of the programe -- "Reinput Times".When you make a mistake in Mode A(Words mode),you must reinput the word some times which it is determined by yourself. <br>
-<br>
-<h3># <h3>
-<br>
-</body>
-</html>
