--- docs/CNAME
@@ -1,1 +1,0 @@
-wordtester.cap
