--- docs/Histories.html
@@ -1,4 +1,0 @@
-# The history of the programe:
-
-# v0.1.x released on Apr 18,2022
-  The first version of the programe!
